+++ client/src/screens/userpanel/Editinvoice.js
@@ -257,7 +257,7 @@ export default function Editinvoice() {
         const { value } = event.target;
         const updatedItems = invoiceData.items.map((item) => {
           if (item.itemId === itemId) {
-            const newQuantity = parseInt(value) >= 0 ? parseInt(value) : 0;
+            const newQuantity = parseFloat(value) >= 0 ? parseFloat(value) : 0;
             const newAmount = calculateDiscountedAmount(item.price, newQuantity, item.discount);
             
             return {
@@ -273,7 +273,7 @@ export default function Editinvoice() {
       };
       
     const onChangeQuantity = (event, itemId) => {
-        let newQuantity = event.target.value ? parseInt(event.target.value) : 1;
+        let newQuantity = event.target.value ? parseFloat(event.target.value) : 1;
         newQuantity = Math.max(newQuantity, 0); // Ensure quantity is not negative
       
         setQuantityMap((prevMap) => ({
